Add frame preview to pod files tab
