+++ front/components/assistant/conversation/interactive_content/FrameRenderer.tsx
@@ -1,5 +1,6 @@
 import { VisualizationActionIframe } from "@app/components/assistant/conversation/actions/VisualizationActionIframe";
 import { useConversationSidePanelContext } from "@app/components/assistant/conversation/ConversationSidePanelContext";
+import { ExportContentDropdown } from "@app/components/assistant/conversation/interactive_content/ExportContentDropdown";
 import { DEFAULT_RIGHT_PANEL_SIZE } from "@app/components/assistant/conversation/constant";
 import { CenteredState } from "@app/components/assistant/conversation/interactive_content/CenteredState";
 import { ShareFrameSheet } from "@app/components/assistant/conversation/interactive_content/frame/ShareFrameSheet";
@@ -9,7 +10,6 @@ import { useDesktopNavigation } from "@app/components/navigation/DesktopNavigati
 import { useVisualizationRevert } from "@app/hooks/conversations";
 import { useHashParam } from "@app/hooks/useHashParams";
 import { useSendNotification } from "@app/hooks/useNotification";
-import config from "@app/lib/api/config";
 import { useAuth } from "@app/lib/auth/AuthContext";
 import { useClientType } from "@app/lib/context/clientType";
 import { clientFetch } from "@app/lib/egress/client";
@@ -19,23 +19,14 @@ import { getErrorFromResponse } from "@app/lib/swr/swr";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
 import { FULL_SCREEN_HASH_PARAM } from "@app/types/conversation_side_panel";
 import type { LightWorkspaceType } from "@app/types/user";
-import { datadogLogs } from "@datadog/browser-logs";
 import {
   ArrowCircleIcon,
-  ArrowDownOnSquareIcon,
   ArrowGoBackIcon,
   Button,
   CheckCircleIcon,
   CloudArrowUpIcon,
   CodeBlock,
   CommandLineIcon,
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
-  DropdownMenuTrigger,
   EyeIcon,
   FullscreenExitIcon,
   FullscreenIcon,
@@ -50,145 +41,6 @@ import React, {
   useRef,
   useState,
 } from "react";
-
-interface ExportContentDropdownProps {
-  iframeRef: React.RefObject<HTMLIFrameElement>;
-  owner: LightWorkspaceType;
-  fileId: string;
-  fileContent: string | null;
-  fileName?: string;
-}
-
-function ExportContentDropdown({
-  iframeRef,
-  owner,
-  fileId,
-  fileContent,
-  fileName,
-}: ExportContentDropdownProps) {
-  const sendNotification = useSendNotification();
-  const [isExportingPdf, setIsExportingPdf] = useState(false);
-
-  const exportAsPng = () => {
-    if (fileContent) {
-      const imgRegex = /<img[^>]+src=["'](https?:\/\/[^"']+)["']/gi;
-      if (imgRegex.test(fileContent)) {
-        sendNotification({
-          type: "error",
-          title: "Cannot export as PNG",
-          description:
-            "Content contains images with external URLs, which are blocked for " +
-            "security purposes. Please use images uploaded to the conversation instead.",
-        });
-        return;
-      }
-    }
-
-    if (iframeRef.current?.contentWindow) {
-      iframeRef.current.contentWindow.postMessage({ type: `EXPORT_PNG` }, "*");
-    } else {
-      datadogLogs.logger.info(
-        "Failed to export content as PNG: No iframe content window found"
-      );
-    }
-  };
-
-  const exportAsPdf = async (orientation: "portrait" | "landscape") => {
-    if (isExportingPdf) {
-      return;
-    }
-
-    setIsExportingPdf(true);
-    try {
-      // Use direct fetch instead of fetcher to avoid JSON parsing of binary PDF data.
-      const response = await clientFetch(
-        `/api/w/${owner.sId}/files/${fileId}/export/pdf`,
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ orientation }),
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error("Failed to generate PDF");
-      }
-
-      // Get the PDF blob and trigger download.
-      const blob = await response.blob();
-      const url = URL.createObjectURL(blob);
-      const link = document.createElement("a");
-      link.href = url;
-
-      link.download = fileName?.replace(/\.[^.]+$/, ".pdf") ?? "frame.pdf";
-
-      link.click();
-      URL.revokeObjectURL(url);
-
-      sendNotification({
-        title: "PDF exported",
-        type: "success",
-        description: "Your PDF has been downloaded.",
-      });
-    } catch (error) {
-      console.error("PDF export failed:", error);
-      sendNotification({
-        title: "PDF Export Failed",
-        type: "error",
-        description: "An error occurred while generating the PDF.",
-      });
-    } finally {
-      setIsExportingPdf(false);
-    }
-  };
-
-  const downloadAsCode = () => {
-    try {
-      const downloadUrl = `${config.getApiBaseUrl()}/api/w/${owner.sId}/files/${fileId}?action=download`;
-      // Open the download URL in a new tab/window. Otherwise we get a CORS error due to the redirection
-      // to cloud storage.
-      window.open(downloadUrl, "_blank");
-    } catch (error) {
-      console.error("Download failed:", error);
-      sendNotification({
-        title: "Download Failed",
-        type: "error",
-        description: "An error occurred while opening the download link.",
-      });
-    }
-  };
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          icon={ArrowDownOnSquareIcon}
-          isSelect
-          label={isExportingPdf ? "Exporting..." : "Export"}
-          variant="ghost"
-          disabled={isExportingPdf}
-        />
-      </DropdownMenuTrigger>
-      <DropdownMenuContent>
-        <DropdownMenuSub>
-          <DropdownMenuSubTrigger disabled={isExportingPdf} label="PDF" />
-          <DropdownMenuSubContent>
-            <DropdownMenuItem onClick={() => exportAsPdf("portrait")}>
-              Portrait
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => exportAsPdf("landscape")}>
-              Landscape
-            </DropdownMenuItem>
-          </DropdownMenuSubContent>
-        </DropdownMenuSub>
-        <DropdownMenuItem onClick={exportAsPng}>PNG</DropdownMenuItem>
-        <DropdownMenuItem onClick={downloadAsCode}>Template</DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
 
 interface FrameRendererProps {
   conversation?: ConversationWithoutContentType;
@@ -425,9 +277,7 @@ export function FrameRenderer({
   if (error) {
     return (
       <div className="flex h-full flex-col">
-        <InteractiveContentHeader
-          onClose={conversation ? onClosePanel : undefined}
-        />
+        <InteractiveContentHeader onClose={onClosePanel} />
         <CenteredState>
           <p className="text-warning-500">
             Error loading file: {error.message}
@@ -439,9 +289,7 @@ export function FrameRenderer({
 
   return (
     <div className="flex h-full flex-col">
-      <InteractiveContentHeader
-        onClose={conversation ? onClosePanel : undefined}
-      >
+      <InteractiveContentHeader onClose={onClosePanel}>
         <div className="flex w-full items-center justify-between">
           <Button
             icon={showCode ? EyeIcon : CommandLineIcon}
